Rename, Create Packages and Minimum update
