+++ TidyBox/WalkingPatternGenerator.cpp
