+++ kinematics/src/hello_world.cpp
@@ -1,0 +1,10 @@
+#include <cstdio>
+
+int main(int argc, char ** argv)
+{
+  (void) argc;
+  (void) argv;
+
+  printf("hello world kinematics package\n");
+  return 0;
+}
