+++ msgs_package/srv/ToKinematicsMessage.msg
@@ -1,0 +1,5 @@
+# R_target
+# R -> Rotarion-Matrixes
+# P_target
+# P -> Joint-Points
+# rad -> Joint-Rotation-Angles
